`unindent.ts` script
Not working very well, but similar problems to the conversion script:
mainly looses empty lines, but also things like a trailing comma in
`foo(3,)` and a few more.  In addition, loses string quotations, which
is unlike the conversion script.

Will try to play with `ts-morph`, committing a checkpoint for this
version.